Add Executor Class
- Fixes failing tests
